# dal-web/backend/openapi/dal-web.openapi.json
{
  "components": {
    "schemas": {
...
      "CurveBuildRunResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dependency_manifest": {
            "items": {
              "$ref": "#/components/schemas/CurveLabDependencyManifestEntryV2"
            },
            "title": "Dependency Manifest",
            "type": "array"
          },
          "diagnostics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostics"
          },
          "draft_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Draft Fingerprint",
            "type": "string"
          },
          "draft_id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Draft Id",
            "type": "string"
          },
          "draft_revision": {
            "minimum": 1.0,
            "title": "Draft Revision",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurveLabErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "pattern": "^[0-9a-f]{32}$",
            "title": "Id",
            "type": "string"
          },
          "native_payload_hash": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Native Payload Hash"
          },
          "parameter_axis": {
            "items": {
              "$ref": "#/components/schemas/ParameterAxisEntryV2"
            },
            "title": "Parameter Axis",
            "type": "array"
          },
          "quote_axis": {
            "items": {
              "$ref": "#/components/schemas/QuoteAxisEntryV2"
            },
            "title": "Quote Axis",
            "type": "array"
          },
          "request": {
            "$ref": "#/components/schemas/CurveDraftDocumentV2"
          },
          "resolved_plan": {
            "additionalProperties": true,
            "title": "Resolved Plan",
            "type": "object"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "state": {
            "enum": [
              "QUEUED",
              "DECLARING",
              "RESOLVING_DEPENDENCIES",
              "NORMALIZING",
              "SOLVING",
              "DIAGNOSTICS",
              "SERIALIZING",
              "SUCCEEDED",
              "FAILED",
              "TIMED_OUT"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "draft_id",
          "draft_revision",
          "draft_fingerprint",
          "state",
          "stale",
          "request",
          "resolved_plan",
          "quote_axis",
          "parameter_axis",
          "dependency_manifest",
          "diagnostics",
          "created_at"
        ],
        "title": "CurveBuildRunResponse",
        "type": "object"
      },
...
      "CurveLabErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/CurveLabErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "CurveLabErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Web UI backend for managing and pricing derivatives portfolios. All quantitative work is delegated to the Derivatives Algorithms Library through its Python public API.",
    "title": "DAL Derivatives Portfolio Management",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/curve-lab/build-runs/{run_id}": {
      "get": {
        "operationId": "get_curve_build_run_api_curve_lab_build_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveBuildRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Build Run",
        "tags": [
          "curve-lab"
        ]
      }
    },
...
    "/api/curve-lab/drafts": {
      "post": {
        "operationId": "create_curve_draft_api_curve_lab_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurveDraftDocumentInputV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Curve Draft",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/drafts/{draft_id}": {
      "get": {
        "operationId": "get_curve_draft_api_curve_lab_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Draft",
        "tags": [
          "curve-lab"
        ]
      },
      "put": {
        "operationId": "update_curve_draft_api_curve_lab_drafts__draft_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurveDraftDocumentInputV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Curve Draft",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/drafts/{draft_id}/build-runs": {
      "post": {
        "operationId": "create_curve_build_run_api_curve_lab_drafts__draft_id__build_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveBuildRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Curve Build Run",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/import-jobs": {
      "post": {
        "operationId": "create_curve_import_job_api_curve_lab_import_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Content-Encoding",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Encoding"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveImportJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Curve Import Job",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/import-jobs/{job_id}": {
      "get": {
        "operationId": "get_curve_import_job_api_curve_lab_import_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveImportJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Import Job",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/quote-canonicalizations": {
      "post": {
        "operationId": "canonicalize_authoring_quote_api_curve_lab_quote_canonicalizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurveLabQuoteCanonicalizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveLabQuoteCanonicalizationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveLabErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Canonicalize Authoring Quote",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/risk-runs": {
      "post": {
        "operationId": "create_curve_risk_run_api_curve_lab_risk_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskRunRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskRunResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Curve Risk Run",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/risk-runs/{run_id}": {
      "get": {
        "operationId": "get_curve_risk_run_api_curve_lab_risk_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskRunResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Risk Run",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/risk-runs/{run_id}/matrices/{matrix_id}": {
      "get": {
        "operationId": "get_curve_risk_matrix_api_curve_lab_risk_runs__run_id__matrices__matrix_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "matrix_id",
            "required": true,
            "schema": {
              "title": "Matrix Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrixResultV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Risk Matrix",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/versions": {
      "get": {
        "operationId": "list_curve_versions_api_curve_lab_versions_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CurveVersionResponse"
                  },
                  "title": "Response List Curve Versions Api Curve Lab Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Curve Versions",
        "tags": [
          "curve-lab"
        ]
      },
      "post": {
        "operationId": "create_curve_version_api_curve_lab_versions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurveVersionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveVersionResponse"
                }
              }
            },
            "description": "Created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Curve Version",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/versions/{version_id}": {
      "get": {
        "operationId": "get_curve_version_api_curve_lab_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Version",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/versions/{version_id}/archive": {
      "post": {
        "operationId": "archive_curve_version_api_curve_lab_versions__version_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Curve Version",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/versions/{version_id}/clone": {
      "post": {
        "operationId": "clone_curve_version_api_curve_lab_versions__version_id__clone_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clone Curve Version",
        "tags": [
          "curve-lab"
        ]
      }
    },
    "/api/curve-lab/versions/{version_id}/native-json": {
      "get": {
        "operationId": "get_curve_version_native_json_api_curve_lab_versions__version_id__native_json_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curve Version Native Json",
        "tags": [
          "curve-lab"
        ]
      }
    },
...
  }
}